refactor(ol-animation-feature): remove unused exported component `ol-animation-feature`
BREAKING CHANGE:
The component `ol-animation-feature` was simply exposing the map, vectorLayer and it's passed properties. There was no need to use this component at all and therefore it has been removed. However, when the component was used in the past, make sure to remove it from your template and move the `ref`'s for map and the layer to the correct components.